Fix Float/Double type mismatch in refreshRawMode()
The real build caught: "Type mismatch: inferred type is Float but
Double? was expected" at plotSafeInset = SQUARE_EDGE_MARGIN_DP * density.
Both operands are Float, and Kotlin doesn't implicitly widen a bare
Float expression to Double on assignment (unlike the adjacent + and <
usages elsewhere in this function, which use Double/Float overloads
that already return Double). Added an explicit .toDouble().

+++ android/app/src/main/java/org/vectair/vcas/car/MainActivity.kt
@@ -1646,7 +1646,7 @@ class MainActivity : Activity() {
             plotHeight = square.squareSize,
             plotOffsetX = square.squareLeft,
             plotOffsetY = square.squareTop,
-            plotSafeInset = SQUARE_EDGE_MARGIN_DP * density,
+            plotSafeInset = (SQUARE_EDGE_MARGIN_DP * density).toDouble(),
             plotBandsNm = activeBandsNm
         )
